Restructure pre-processing of data
This commit cleans up the previous pre-processing steps.

The fundamental issue with the previous approach was: the target accounts are defined by combining multiple lists of accounts, but there was no way to track multiple membership in lists. So, for example, it was difficult to subset the resulting elites file into _just_ the 116th Congress or _just_ the 117th Congress. To get around this, we need a source file that defines membership in each list as its own column.

This also corrects the display name vs. screen name issue in the presidents and covid_elites lists.